+++ types/beep.d.ts
@@ -1,5 +1,5 @@
 
-declare module "@rollup/plugin-beep" {
+declare module '@rollup/plugin-beep' {
 
   export default function beep(): void;