+++ changelog.md
@@ -1,5 +1,26 @@
 ## _CHANGELOG_
 
+## v1.1.0 - 09/09/21
+
+- Updated readme
+- Reverted @rollup/plugin-typescript back to rollup-plugin-typescript2
+- Added rollup-plugin-postcss
+- Added rollup-plugin-scss
+- Removed rollup-plugin-ts
+- Made rollup an optional dependency
+- Added the following optional dependencies
+  - autoprefixer
+  - typescript
+  - tsconfig
+  - postcss
+  - node-sass
+
+## v1.0.1 - 03/00/21
+
+- Replaced rollup-plugin-typescript2 with @rollup/plugin-typescript
+- Added @rollup/plugin-html
+- Added linked JSDoc annotations
+
 ## v1.0.0 - 23/08/21
 
 - Initial Release